+++ fault-tolerance/src/main/java/io/helidon/faulttolerance/Fallback.java
@@ -28,7 +28,7 @@ import io.helidon.common.reactive.Multi;
 import io.helidon.common.reactive.Single;
 
 /**
- * Fallback allows the user to execute an alternative supplier of results in case the usual one fails.
+ * Fallback allows the user to execute an alternative function in case the provided supplier fails.
  * <p>
  * In case you call the {@link #invokeMulti(java.util.function.Supplier)} method, the following restriction applies:
  * <ul>
@@ -122,7 +122,8 @@ public interface Fallback<T> extends FtHandlerTyped<T> {
          * @param classes classes to fallback on
          * @return updated builder instance
          */
-        public Builder<T> applyOn(Class<? extends Throwable>... classes) {
+        @SafeVarargs
+        public final Builder<T> applyOn(Class<? extends Throwable>... classes) {
             applyOn.clear();
             Arrays.stream(classes)
                     .forEach(this::addApplyOn);
@@ -148,7 +149,8 @@ public interface Fallback<T> extends FtHandlerTyped<T> {
          * @param classes classes not to fallback on
          * @return updated builder instance
          */
-        public Builder<T> skipOn(Class<? extends Throwable>... classes) {
+        @SafeVarargs
+        public final Builder<T> skipOn(Class<? extends Throwable>... classes) {
             skipOn.clear();
             Arrays.stream(classes)
                     .forEach(this::addSkipOn);